+++ src/format.rs
@@ -1,16 +1,10 @@
-use std::fmt::Display;
-
-use chumsky::prelude::Simple;
-
-use crate::Token;
-
-pub type Id = u32;
-
-pub type StateConjunction = Vec<Id>;
+pub type StateConjunction = Vec<crate::Id>;
 
 pub type AtomicProposition = String;
 
 pub type AliasName = String;
+
+pub type AcceptanceAtom = (bool, crate::Id);
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum LabelExpression {
@@ -21,26 +15,6 @@ pub enum LabelExpression {
     And(Vec<LabelExpression>),
     Or(Vec<LabelExpression>),
 }
-
-impl LabelExpression {
-    pub fn boolean(b: bool) -> Self {
-        Self::Boolean(b)
-    }
-
-    pub fn alias<I: Display>(name: I) -> Self {
-        Self::Alias(name.to_string())
-    }
-
-    pub fn not(expr: Self) -> Self {
-        Self::Not(Box::new(expr))
-    }
-
-    pub fn int(i: u32) -> Self {
-        Self::Integer(i)
-    }
-}
-
-pub type AcceptanceAtom = (bool, Id);
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum AcceptanceCondition {
@@ -137,32 +111,6 @@ impl TryFrom<String> for Property {
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum AcceptanceInfo {
-    Int(Id),
+    Int(crate::Id),
     Identifier(String),
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub enum HeaderItem {
-    States(Id),
-    Start(StateConjunction),
-    AP(Vec<AtomicProposition>),
-    Alias(AliasName, LabelExpression),
-    Acceptance(Id, AcceptanceCondition),
-    AcceptanceName(AcceptanceName, Vec<AcceptanceInfo>),
-    // Correspond to tool name and optional version number
-    Tool(String, Option<String>),
-    Name(String),
-    Properties(Vec<Property>),
-}
-
-#[derive(Debug)]
-pub struct HoaAutomaton {
-    version: String,
-    headers: Vec<HeaderItem>,
-}
-
-impl HoaAutomaton {
-    pub fn new((version, headers): (String, Vec<HeaderItem>)) -> Self {
-        Self { version, headers }
-    }
 }
